Migrations fix

+++ training-net/src/Migrations/20180823131401_InitialMigration.cs
@@ -17,7 +17,7 @@ namespace trainingnet.Migrations
                     Title = table.Column<string>(nullable: true),
                     ReleaseDate = table.Column<DateTime>(nullable: false),
                     Genre = table.Column<string>(nullable: true),
-                    Price = table.Column<decimal>(nullable: false)
+                    Price = table.Column<float>(nullable: false)
                 },
                 constraints: table =>
                 {
